:lipstick:

+++ src/stdlib/oniguruma/src/onig-reg-exp.h
@@ -13,11 +13,11 @@ class OnigRegExp {
 
     bool Contains(const std::string& value);
     int LocationAt(int index);
-    void Search(const std::string &searchString, int position, OnigResult **result);
+    void Search(const std::string &searchString, size_t position, OnigResult **result);
 
   private:
     std::string source_;
-    regex_t *regex_;
+    regex_t* regex_;
 };
 
 #endif // ONIG_REG_EXP_H_
